services/requirements.txt
Flask==2.3.2
boto3==1.28.32
python-dotenv==1.0.0